refactor(infra): PerplexityをResearch構成へ再編（Search→Research、バレル化）\n\n- external/perplexity/Research/index.ts に実装移動（PerplexityResearchClient）\n- index.ts は Research を再エクスポート\n- 互換エイリアス PerplexityClient を維持\n- テストを Research/index.test.ts へ移動\n\nRefs: #9
