+++ src/shared/infrastructure/external/perplexity/Search/index.ts
@@ -1,0 +1,143 @@
+/**
+ * Perplexity API Infrastructure - Search
+ * 外部API（Perplexity）との通信（検索/リサーチ）を担当
+ */
+
+import OpenAI from "openai";
+import type { VoicePattern } from "@/shared/api/generated/models/voicePattern";
+
+export interface PerplexityConfig {
+  apiKey: string;
+  model?: string;
+  baseUrl?: string;
+}
+
+export interface ResearchContext {
+  query: string;
+  selectedText?: string;
+  voiceCommand?: VoicePattern;
+  researchId?: string;
+}
+
+export interface PerplexitySearchResult {
+  title: string;
+  url: string;
+  snippet: string;
+  date?: string;
+}
+
+export interface PerplexityResponse extends OpenAI.Chat.ChatCompletion {
+  citations?: string[];
+  search_results?: PerplexitySearchResult[];
+  related_questions?: string[];
+}
+
+export class PerplexityAPIError extends Error {
+  constructor(
+    message: string,
+    public readonly code?: string,
+    public readonly status?: number,
+  ) {
+    super(message);
+    this.name = "PerplexityAPIError";
+  }
+}
+
+export interface IResearchAPIRepository {
+  search(context: ResearchContext): Promise<PerplexityResponse>;
+}
+
+export class PerplexitySearchClient implements IResearchAPIRepository {
+  private readonly client: OpenAI;
+  private readonly model: string;
+
+  constructor(config: PerplexityConfig) {
+    if (!config.apiKey?.trim()) {
+      throw new PerplexityAPIError("API key is required");
+    }
+
+    this.client = new OpenAI({
+      apiKey: config.apiKey,
+      baseURL: config.baseUrl || "https://api.perplexity.ai",
+    });
+
+    this.model = config.model || "sonar";
+  }
+
+  async search(context: ResearchContext): Promise<PerplexityResponse> {
+    if (!context.query?.trim()) {
+      throw new PerplexityAPIError("Query is required");
+    }
+
+    try {
+      const messages = this.buildPrompt(context);
+
+      const response = await this.client.chat.completions.create({
+        model: this.model,
+        messages,
+        max_tokens: 2000,
+        temperature: 0.2,
+        return_citations: true,
+        return_related_questions: true,
+        stream: false,
+      } as OpenAI.Chat.ChatCompletionCreateParams & {
+        return_citations: boolean;
+        return_related_questions: boolean;
+      });
+
+      return response as PerplexityResponse;
+    } catch (error) {
+      if (error instanceof OpenAI.APIError) {
+        throw new PerplexityAPIError(
+          `Perplexity API error: ${error.message}`,
+          error.code || undefined,
+          error.status,
+        );
+      }
+      throw new PerplexityAPIError(
+        `Network error: ${error instanceof Error ? error.message : "Unknown error"}`,
+      );
+    }
+  }
+
+  // プロンプト構築
+  private buildPrompt(
+    context: ResearchContext,
+  ): OpenAI.Chat.ChatCompletionMessageParam[] {
+    const systemPrompt = `あなたは世界最高レベルのリサーチ専門家です。以下の原則に従って包括的なリサーチを実施してください：
+
+【リサーチ原則】
+1. 最新の情報を基に正確で信頼性の高い回答を提供
+2. 出典の明記と引用の適切な付与
+3. ユーザー意図に忠実で過剰な拡張を避ける
+4. 重要なポイントは見出し付きで構造化
+5. 関連する追加の調査観点を提案
+
+【出力形式】
+- 見出しを使った構造化された回答
+- 重要な情報には適切な引用を含める
+- 関連する質問や調査観点を提案`;
+
+    const messages: OpenAI.Chat.ChatCompletionMessageParam[] = [
+      { role: "system", content: systemPrompt },
+    ];
+
+    if (context.selectedText) {
+      messages.push({
+        role: "user",
+        content: `【選択されたテキスト】\n${context.selectedText}\n\n【リサーチクエリ】\n${context.query}`,
+      });
+    } else {
+      messages.push({ role: "user", content: context.query });
+    }
+
+    return messages;
+  }
+}
+
+// 型の再エクスポート
+export { OpenAI };
+export type PerplexityMessage = OpenAI.Chat.ChatCompletionMessageParam;
+
+// 互換エイリアス（旧名）
+export { PerplexitySearchClient as PerplexityClient };
